--- app/src/main/java/beepbeep/pixelsforreddit/PixelApplication.kt
@@ -1,18 +1,0 @@
-package beepbeep.pixelsforreddit
-
-import android.app.Application
-import android.content.Context
-import com.facebook.stetho.Stetho
-
-class PixelApplication : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-        _applicationContext = this
-        Stetho.initializeWithDefaults(this)
-    }
-
-    companion object {
-        private var _applicationContext: Context? = null
-    }
-}
