--- app/src/main/java/beepbeep/pixelsforreddit/extension/BooleanExtension.kt
@@ -1,10 +1,0 @@
-package beepbeep.pixelsforreddit.extension
-
-import android.view.View
-
-fun Boolean.visibility(): Int {
-    if (this) {
-        return View.VISIBLE
-    }
-    return View.GONE
-}
