--- app/src/main/java/beepbeep/pixelsforreddit/extension/UnixTimeToRelativeStringExtension.kt
@@ -1,10 +1,0 @@
-package beepbeep.pixelsforreddit.extension
-
-import android.text.format.DateUtils
-import java.util.*
-
-fun Int.toRelativeTimeString(): String {
-    val oldUnixTime = this.toLong() * 1000
-    val nowUnixTime = Date().time
-    return DateUtils.getRelativeTimeSpanString(oldUnixTime, nowUnixTime, DateUtils.MINUTE_IN_MILLIS).toString()
-}
