--- app/src/main/java/beepbeep/pixelsforreddit/extension/UrlStringExtension.kt
@@ -1,8 +1,0 @@
-package beepbeep.pixelsforreddit.extension
-
-/**
- * https://stackoverflow.com/a/169631/75579
- */
-fun String.isImageUrl(): Boolean {
-    return "(?:([^:/?#]+):)?(?://([^/?#]*))?([^?#]*\\.(?:jpg|gif|png))(?:\\?([^#]*))?(?:#(.*))?".toRegex().matches(this)
-}
