--- app/src/main/java/beepbeep/pixelsforreddit/home/HomeRepo.kt
@@ -1,30 +1,0 @@
-package beepbeep.pixelsforreddit.home
-
-import android.content.Context
-import beepbeep.pixelsforreddit.preference.RedditPreference
-import com.worker8.redditapi.RedditApi
-import io.reactivex.Scheduler
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
-
-class HomeRepo(val context: Context, subreddit: String = "pics") {
-    var redditApi: RedditApi
-
-    init {
-        redditApi = RedditApi(subreddit)
-    }
-
-    fun getMorePosts() = redditApi.getMorePosts()
-    fun getMainThread(): Scheduler = AndroidSchedulers.mainThread()
-    fun getBackgroundThread(): Scheduler = Schedulers.io()
-
-    fun selectSubreddit(subreddit: String) {
-        redditApi = RedditApi(subreddit)
-    }
-
-    fun saveSubredditSharedPreference(subreddit: String) =
-        RedditPreference.saveSelectedSubreddit(context, subreddit)
-
-    fun getSubredditSharedPreference() =
-        RedditPreference.getSelectedSubreddit(context)
-}
