+++ app/src/main/java/beepbeep/pixelsforredditx/PixelApplication.kt
@@ -1,0 +1,18 @@
+package beepbeep.pixelsforredditx
+
+import android.app.Application
+import android.content.Context
+import com.facebook.stetho.Stetho
+
+class PixelApplication : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        _applicationContext = this
+        Stetho.initializeWithDefaults(this)
+    }
+
+    companion object {
+        private var _applicationContext: Context? = null
+    }
+}
