+++ app/src/main/java/beepbeep/pixelsforredditx/common/NetworkState.kt
@@ -1,0 +1,18 @@
+package beepbeep.pixelsforredditx.common
+
+enum class Status {
+    LOADING,
+    SUCCESS,
+    FAILED
+}
+
+@Suppress("DataClassPrivateConstructor")
+data class NetworkState private constructor(
+    val status: Status,
+    val msg: String? = null) {
+    companion object {
+        val SUCCESS = NetworkState(Status.SUCCESS)
+        val LOADING = NetworkState(Status.LOADING)
+        fun error(msg: String?) = NetworkState(Status.FAILED, msg)
+    }
+}
