+++ app/src/main/java/beepbeep/pixelsforredditx/common/glideOptions/MyGlideModule.kt
@@ -1,0 +1,22 @@
+package beepbeep.pixelsforredditx.common.glideOptions
+
+import android.content.Context
+import android.graphics.Bitmap
+import android.graphics.drawable.Drawable
+import com.bumptech.glide.GlideBuilder
+import com.bumptech.glide.annotation.GlideModule
+import com.bumptech.glide.load.resource.bitmap.BitmapTransitionOptions
+import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
+import com.bumptech.glide.module.AppGlideModule
+
+@GlideModule
+class MyGlideModule : AppGlideModule() {
+    override fun applyOptions(context: Context, builder: GlideBuilder) {
+        builder.setDefaultTransitionOptions(Drawable::class.java, DrawableTransitionOptions.withCrossFade())
+            .setDefaultTransitionOptions(Bitmap::class.java, BitmapTransitionOptions.withCrossFade())
+    }
+
+    override fun isManifestParsingEnabled(): Boolean {
+        return false
+    }
+}
