+++ app/src/main/java/beepbeep/pixelsforredditx/extension/BooleanExtension.kt
@@ -1,0 +1,10 @@
+package beepbeep.pixelsforredditx.extension
+
+import android.view.View
+
+fun Boolean.visibility(): Int {
+    if (this) {
+        return View.VISIBLE
+    }
+    return View.GONE
+}
