+++ app/src/main/java/beepbeep/pixelsforredditx/extension/ViewExtension.kt
@@ -1,0 +1,13 @@
+package beepbeep.pixelsforredditx.extension
+
+import android.view.View
+
+fun View.dpWidth(): Int {
+    val density = resources.displayMetrics.density
+    return (width / density).toInt()
+}
+
+fun View.dpHeight(): Int {
+    val density = resources.displayMetrics.density
+    return (height / density).toInt()
+}
